requirements.txt
python-twitter
requests[security]
